Implemented 1) Single type injection 2) multiple types injections WIP
