--- src/main/java/WordApplication.java
@@ -1,8 +1,0 @@
-
-public class WordApplication {
-
-    public static void main(String[] args) {
-        //Do nothing until the task is done
-    }
-
-}
