--- src/main/java/dao/Constants.java
@@ -1,5 +1,0 @@
-package dao;
-
-public class Constants {
-    public static final String BASE_DIRECTORY = "Users";
-}
