+++ src/main/java/tech/frost/SOMEINT.java
@@ -1,0 +1,6 @@
+package tech.frost;
+
+public interface SOMEINT {
+
+    void print();
+}
