+++ src/main/java/tech/frost/TestBeanDependency.java
@@ -1,0 +1,13 @@
+package tech.frost;
+
+public class TestBeanDependency implements SOMEINT {
+
+    public TestBeanDependency() {
+        System.out.println("beanDepCreated");
+    }
+
+    @Override
+    public void print() {
+        System.out.println("message");
+    }
+}
