+++ src/main/java/tech/frost/configuration/configurator/ObjectConfigurator.java
@@ -1,0 +1,7 @@
+package tech.frost.configuration.configurator;
+
+import tech.frost.configuration.infrastucture.ApplicationContext;
+
+public interface ObjectConfigurator {
+    void configure(Object t, ApplicationContext context);
+}
