+++ src/main/java/tech/frost/configuration/configurator/ProxyObjectConfigurator.java
@@ -1,0 +1,7 @@
+package tech.frost.configuration.configurator;
+
+import tech.frost.configuration.infrastucture.ApplicationContext;
+
+public interface ProxyObjectConfigurator {
+    <T> T configure(T t, ApplicationContext context);
+}
