+++ src/main/java/tech/frost/configuration/configurator/interfaces/InjectProperty.java
@@ -1,0 +1,13 @@
+package tech.frost.configuration.configurator.interfaces;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.FIELD;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Retention(RUNTIME)
+@Target(FIELD)
+public @interface InjectProperty {
+    String value() default "";
+}
