+++ src/main/java/tech/frost/dao/Constants.java
@@ -1,0 +1,5 @@
+package tech.frost.dao;
+
+public class Constants {
+    public static final String BASE_DIRECTORY = "Users";
+}
