+++ src/main/java/tech/frost/dao/UserEntity.java
@@ -1,0 +1,20 @@
+package tech.frost.dao;
+
+public class UserEntity {
+    private final String login;
+    private final String pass;
+
+
+    public UserEntity(String login, String pass) {
+        this.login = login;
+        this.pass = pass;
+    }
+
+    public String getLogin() {
+        return login;
+    }
+
+    public String getPass() {
+        return pass;
+    }
+}
