+++ src/main/resources/application.properties
@@ -1,0 +1,1 @@
+value=testValue
